(83) refactor model folder-2
